pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telemetry-platform"
version = "1.0.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.4",
    "arq>=0.25.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "alembic>=1.13.1",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
    "types-redis>=4.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "W", "F", "I", "B", "C90", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["app/db/migrations/*"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
